+++ src/lib/index.js
@@ -1,5 +1,6 @@
 /* @flow */
 
+export * from "./api";
 export * from "./errors";
 export * from "./isRTLLanguage";
 export * from "./security";